Split format and check rules into separate packages
