+++ src/python/pants/backend/experimental/python/lint/ruff/check/register.py
@@ -6,9 +6,15 @@
 See https://www.pantsbuild.org/docs/python-linters-and-formatters and https://docs.astral.sh/ruff/
 """
 
-from pants.backend.python.lint.ruff import check_rules as ruff_rules
 from pants.backend.python.lint.ruff import skip_field, subsystem
+from pants.backend.python.lint.ruff.check import rules as ruff_rules
+from pants.backend.python.lint.ruff.check import skip_field as ruff_check_skip_field
 
 
 def rules():
-    return (*ruff_rules.rules(), *skip_field.rules(), *subsystem.rules())
+    return (
+        *ruff_rules.rules(),
+        *skip_field.rules(),
+        *ruff_check_skip_field.rules(),
+        *subsystem.rules(),
+    )
