+++ src/python/pants/backend/experimental/python/lint/ruff/format/register.py
@@ -6,9 +6,15 @@
 See https://www.pantsbuild.org/docs/python-linters-and-formatters and https://docs.astral.sh/ruff/
 """
 
-from pants.backend.python.lint.ruff import fmt_rules as ruff_fmt_rules
 from pants.backend.python.lint.ruff import skip_field, subsystem
+from pants.backend.python.lint.ruff.format import rules as ruff_fmt_rules
+from pants.backend.python.lint.ruff.format import skip_field as ruff_format_skip_field
 
 
 def rules():
-    return (*ruff_fmt_rules.rules(), *skip_field.rules(), *subsystem.rules())
+    return (
+        *ruff_fmt_rules.rules(),
+        *skip_field.rules(),
+        *ruff_format_skip_field.rules(),
+        *subsystem.rules(),
+    )
