+++ src/python/pants/backend/python/lint/ruff/format/rules.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import Optional
 
 from pants.backend.python.lint.ruff.common import RunRuffRequest, run_ruff
-from pants.backend.python.lint.ruff.skip_field import SkipRuffFormatField
+from pants.backend.python.lint.ruff.format.skip_field import SkipRuffFormatField
 from pants.backend.python.lint.ruff.subsystem import Ruff, RuffMode
 from pants.backend.python.target_types import (
     InterpreterConstraintsField,
