+++ src/python/pants/backend/python/lint/ruff/rules_integration_test.py
@@ -7,16 +7,20 @@ from pathlib import Path
 
 import pytest
 
+import pants.backend.python.lint.ruff.check.skip_field
+import pants.backend.python.lint.ruff.format.skip_field
 from pants.backend.python import target_types_rules
 from pants.backend.python.lint.ruff import skip_field
-from pants.backend.python.lint.ruff.check_rules import (
+from pants.backend.python.lint.ruff.check import skip_field as ruff_check_skip_field
+from pants.backend.python.lint.ruff.check.rules import (
     RuffCheckFieldSet,
     RuffFixRequest,
     RuffLintRequest,
 )
-from pants.backend.python.lint.ruff.check_rules import rules as ruff_rules
-from pants.backend.python.lint.ruff.fmt_rules import RuffFormatFieldSet, RuffFormatRequest
-from pants.backend.python.lint.ruff.fmt_rules import rules as ruff_fmt_rules
+from pants.backend.python.lint.ruff.check.rules import rules as ruff_check_rules
+from pants.backend.python.lint.ruff.format import skip_field as ruff_format_skip_field
+from pants.backend.python.lint.ruff.format.rules import RuffFormatFieldSet, RuffFormatRequest
+from pants.backend.python.lint.ruff.format.rules import rules as ruff_fmt_rules
 from pants.backend.python.lint.ruff.subsystem import rules as ruff_subsystem_rules
 from pants.backend.python.target_types import PythonSourcesGeneratorTarget
 from pants.core.goals.fix import FixResult
@@ -39,9 +43,11 @@ UNFORMATTED_FILE = 'a ="string without any placeholders"\n'
 def rule_runner() -> RuleRunner:
     return RuleRunner(
         rules=[
-            *ruff_rules(),
+            *ruff_check_rules(),
             *ruff_fmt_rules(),
             *skip_field.rules(),
+            *ruff_check_skip_field.rules(),
+            *ruff_format_skip_field.rules(),
             *ruff_subsystem_rules(),
             *config_files.rules(),
             *target_types_rules.rules(),
@@ -187,7 +193,10 @@ def test_skip_check_field(rule_runner: RuleRunner) -> None:
         rule_runner.get_target(Address("", target_name="t", relative_file_path="unformatted.py")),
     ]
     for tgt in tgts:
-        assert tgt.get(skip_field.SkipRuffCheckField).value is True
+        assert (
+            tgt.get(pants.backend.python.lint.ruff.check.skip_field.SkipRuffCheckField).value
+            is True
+        )
 
     fix_result, lint_result, fmt_result = run_ruff(rule_runner, tgts)
 
@@ -215,7 +224,10 @@ def test_skip_format_field(rule_runner: RuleRunner) -> None:
         rule_runner.get_target(Address("", target_name="t", relative_file_path="unformatted.py")),
     ]
     for tgt in tgts:
-        assert tgt.get(skip_field.SkipRuffFormatField).value is True
+        assert (
+            tgt.get(pants.backend.python.lint.ruff.format.skip_field.SkipRuffFormatField).value
+            is True
+        )
 
     fix_result, lint_result, fmt_result = run_ruff(rule_runner, tgts)
